+++ logic/sector.cpp
@@ -114,10 +114,10 @@ const GnfrSector& EmissionSector::gnfr_sector() const noexcept
     return get_nfr_sector().gnfr();
 }
 
-int64_t EmissionSector::id() const noexcept
+int32_t EmissionSector::id() const noexcept
 {
     return std::visit([](auto& sectorType) {
-        return static_cast<int64_t>(sectorType.id());
+        return static_cast<int32_t>(sectorType.id());
     },
                       _sector);
 }
